requirements.txt
Flask
Flask-CORS
openai
anthropic
httpx
gunicorn
orjson
torch
mace-torch
ase
matplotlib
paramiko
pyscf
dftd3
dftd4
scikit-image
cuequivariance
gpu4pyscf-cuda12x